Implement ivar and vvar
These are two complements to the traditional var, where the system links value paths pointing
to them to the getter and setter counterpart methods.

The names of the getters and setters are constructed by prepending "get" or "set" and
converting the first letter of the var name to uppercase, effectively making use of the
camelCase naming convention. E.g: foo makes calls to getFoo() and setFoo()

The difference between the two is that ivars generate the necessary return statements and
assign statements for the getter and the setter respectively, and the vvar require the
developers to write those statements themselves.