Make function calls contain the name of the thing they call
This reverts from the notion that everything is always wrapped in a ValuePath, since it
was generating too much bloat in the AST. A function call can now exist as a standalone
node, while still being able to exist inside a value path. In the latter case, it always
points to a method in a class.