Implement new multi-stage type guessing
The way it was done before would lead to an infinite loop when using nested expressions,
since the code was searching up the node tree and looking inside the child nodes at the
same time. Now stuff gets done in a specific order, first trickling down the types we know
for sure, then trying to guess the remaining nodes.