Implement automatic copying of resources
Usage: mark the field of the class that is used as a definition for a rule with the #resource instruction,
then at the end of parsing everything the rules will be searched for the string literal containing the name
of the file to be copied. After packaging, the resources are copied using the cp command.